py/objobject: Fix object.__new__ for metaclass-created classes.
The object.__new__ function was using mp_obj_is_type() which only
checks if the object is exactly of type 'type', not if it's a
subclass. This caused failures for classes created by custom
metaclasses. Changed to use mp_obj_is_subclass_fast() to check if
the metaclass is a subclass of type, allowing object.__new__ to
work with enum classes and other metaclass-created classes.

Signed-off-by: Andrew Leech <andrew.leech@planetinnovation.com.au>