py/objobject: Fix object.__new__ segfault with non-type arguments.
The object.__new__ function was calling mp_obj_is_subclass_fast() without
first verifying that the argument was a type object. When called with
non-type arguments like integers, this caused a segmentation fault.

Changed to first check if the argument is a type object (by checking if
its metaclass is type or a subclass of type) before performing further
validation.

Signed-off-by: Andrew Leech <andrew.leech@planetinnovation.com.au>